code/embedding-function/requirements.txt
azure-functions==1.21.3
python-dotenv==1.0.1
azure-ai-formrecognizer==3.3.3
azure-storage-blob==12.24.0
azure-identity==1.19.0
flask[async]==3.0.3
openai==1.61.0
langchain==0.2.17
langchain-community==0.2.19
langchain-openai==0.1.25
requests==2.32.3
tiktoken==0.8.0
azure-storage-queue==12.12.0
beautifulsoup4==4.12.3
chardet==5.2.0
azure-search-documents==11.6.0b1
azure-ai-contentsafety==1.0.0
python-docx==1.1.2
azure-keyvault-secrets==4.9.0
pandas==2.2.3
azure-monitor-opentelemetry>=1.6.4
opentelemetry-instrumentation-httpx>=0.50b0
pillow==10.4.0
azure-mgmt-cognitiveservices>=13.6.0
jsonschema>=4.23.0
pydantic==2.7.4
azure-ai-ml>=1.24.0
